+++ packages/connectivity_status/test/helpers/connectivity_mocks.dart
@@ -1,3 +1,3 @@
 export 'fake_connectivity_metered_data_source.dart';
+export 'fake_connectivity_repository.dart';
 export 'mock_connectivity_data_source.dart';
-export 'mock_connectivity_repository.dart';